parasite: Introduce infect_ctx
This routine will become the part of compel A[PB]I and
will be given to user as configuration structure.

For now put there the pointer on a socket living in victim's
netns. Since CRIU creates such and generic compel user would
want compel to do it for him, this socket should be the
infect_ctx part.

Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>